requirements.txt
bitstring==3.1.5
certifi==2017.11.5
chardet==3.0.4
click==6.7
forecastiopy==0.22
google-api-python-client==1.6.4
httplib2==0.10.3
hupper==1.0
idna==2.6
ipaddress==1.0.19
Jinja2==2.10
keyring==8.7
keyrings.alt==1.3
lifehud==0.0
Mako==1.0.7
MarkupSafe==1.0
oauth2client==4.1.2
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pyicloud==0.9.1
pyramid==1.9.1
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pytz==2017.3
PyYAML==3.12
repoze.lru==0.7
requests==2.18.4
rsa==3.4.2
six==1.11.0
translationstring==1.3
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.0rc1
zope.deprecation==4.3.0
zope.interface==4.4.3
